total refactor of CalendarPlugin
